+++ addon/utils/reload-local-records.js
@@ -59,18 +59,14 @@ export function reloadLocalRecords(type, reload, projectionName, params) {
 export function createLocalRecord(store, localAdapter, localStore, modelType, record, projection, params) {
   let _this = this;
   let dexieService = Ember.getOwner(store).lookup('service:dexie');
-  let unloadRecordFromStore = () => {
-    if (params && params.unloadSyncedRecords) {
-      if (store.get('onlineStore')) {
-        store.get('onlineStore').unloadRecord(record);
-      } else {
-        store.unloadRecord(record);
-      }
-    }
-  };
+  if (params && params.unloadSyncedRecords) {
+    _this.get('_recordsToUnload').push(record);
+  }
 
   dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') + 1);
   if (record.get('id')) {
+    // TODO: if record from online store should not be unloaded then it will have dirty attribute.
+    record.set('syncDownTime', new Date());
     var snapshot = record._createSnapshot();
     let fieldsToUpdate = projection ? projection.attributes : null;
     return new Ember.RSVP.Promise((resolve, reject) => localAdapter.addHashForBulkUpdateOrCreate(localStore, modelType, snapshot, fieldsToUpdate).then(() => {
@@ -87,7 +83,7 @@ export function createLocalRecord(store, localAdapter, localStore, modelType, re
     }).catch((reason) => {
       Ember.Logger.error(reason);
       reject(reason);
-    }).finally(unloadRecordFromStore));
+    }));
   } else {
     var recordName = record.constructor && record.constructor.modelName;
     var warnMessage = 'Record ' + recordName + ' does not have an id, therefor we can not create it locally: ';
@@ -117,7 +113,7 @@ function createLocalRecords(store, localAdapter, localStore, modelType, records,
         }
       }, reject));
   });
-  return RSVP.all(createdRecordsPromises);
+  return RSVP.all(createdRecordsPromises).then(() => _this._unloadRecordsAfterSyncDown(store, params));
 }
 
 export function syncDownRelatedRecords(store, mainRecord, localAdapter, localStore, projection, params) {
@@ -165,11 +161,7 @@ export function syncDownRelatedRecords(store, mainRecord, localAdapter, localSto
         } else {
           if (isEmbedded(store, modelType, belongToName)) {
             var relatedRecord = mainRecord.get(belongToName);
-            if (!Ember.isNone(relatedRecord)) {
-              promises.pushObject(
-                createRelatedBelongsToRecord(store, relatedRecord, localAdapter, localStore, Ember.isNone(attrs) ? null : attrs[belongToName])
-              );
-            }
+            createRelatedBelongsToRecordFunction(relatedRecord);
           }
         }
       }
@@ -189,7 +181,7 @@ export function syncDownRelatedRecords(store, mainRecord, localAdapter, localSto
         } else {
           if (isEmbedded(store, modelType, hasManyName)) {
             var relatedRecords = mainRecord.get(hasManyName);
-            promises.pushObjects(createRelatedHasManyRecords(store, relatedRecords, localAdapter, localStore, Ember.isNone(attrs) ? null : attrs[hasManyName]));
+            createRelatedHasManyRecordsFunction(relatedRecords);
           }
         }
       }